requirements.txt
Flask>=1.0.3
PyJWT>=2.0.0
Werkzeug>=1.0.1
blinker>=1.4
pytest>=6.2.1
pypandoc>=1.5
ruamel.yaml>=0.16.12